Expose Ripgrep's support for multiple search paths
    rg [OPTIONS] PATTERN [PATH ...]

Make the directory concept explicit, where before it was implicit:

- Update `deadgrep--arguments' to take a list of search paths, rather
than assuming that "." is correct.

- Decouple the search buffer's "current directory" concept from "paths
to search".  Instead, establish a convention that the first path to
search is the search buffer's current directory.  Users are
responsible for providing correct absolute or relative paths.

For backwards compatibility, allow `deadgrep-project-root-function' to
return a string or a list of strings.

Save the search paths in new local variable `deadgrep--search-paths',
so refreshes work as expected.